Restructuring
Complete restructure.

The server now only contains one ServerController, which it delegates
tasks to.

The ServerController class is responsible for creating and having an
overview of the GameControllers and finding the right GameController for
an incoming socket. It is also responsible for removing GameControllers
whenever their games are done.

GameControllers represent a single game and implement a
Publish/Subscribe pattern to send out data to sockets. Contrary to the
old version, the GameController class maintains a cache of the data,
making sure that no data is fetched twice, and sending data immediatly
if it has already been fetched.